CRUD AND LOGIN

+++ Building-Database-Applications-in-PHP/login.php
@@ -1,13 +1,6 @@
 <?php
-require_once "SQl-Queries.php";
-// Show THe Members
-echo "<pre>\n";
-$sqlShow = "SELECT * FROM Usuarios";
-$stmt_show = $pdo->query($sqlShow);
-while($row = $stmt_show->fetch(PDO::FETCH_ASSOC)){
-       echo "<form method='post'><table><tr><td>{$row["user_id"]}</td> <td>{$row["name"]}</td><td>{$row["email"]}</td></tr></table> <input type='hidden' name='toDelete' value={$row['user_id']}/> <input type='submit' value='DELETE'></form>";
-}
-echo "</pre>\n";
+require_once "./SQl-Queries.php";
+
 
 // Add to database
 if(isset($_POST["name"]) && isset($_POST["pw"]) && isset($_POST["email"])) {
@@ -16,11 +9,11 @@ if(isset($_POST["name"]) && isset($_POST["pw"]) && isset($_POST["email"])) {
     $pw = $_POST["pw"];
     // Placeholder to avoid injection :placeholder
     $sql = "INSERT INTO Usuarios (name,email,pw) VALUES (:name,:email, :pw)";
-    $pdo->setAttribute(PDO::ATTR_ERRMODE,PDO::ERRMODE_EXCEPTION);
+
     try {
         $stmt = $pdo->prepare($sql);
         $stmt->execute([":name" => $name, ":email" => $email, ":pw" => $pw]);
-       // Try Cath only show the error and no secret information
+       // Try Catch only show the error and no secret information
     }catch (Exception $ex){
         echo("Internal Error, Please contact support");
         error_log("Exception message: ".$ex->getMessage());
@@ -28,16 +21,19 @@ if(isset($_POST["name"]) && isset($_POST["pw"]) && isset($_POST["email"])) {
     }
 
 }
+if(isset($_POST['toDelete'])){
 
-// Delete from database
-// Is not necesary use isset[$_POST[toDelete] because the user is in the table already
-$stmtDelete = "DELETE FROM `Usuarios` WHERE user_id=(:name)";
+    // Delete from database
+// Is not necessary use isset[$_POST[toDelete] because the user is in the table already
+    $stmtDelete = "DELETE FROM `Usuarios` WHERE user_id=(:name)";
 // Make the User_id a var
-$userToDelete= $_POST["toDelete"];
+    $userToDelete= $_POST['toDelete'];
 // Prepare the statement
-$stmtDelete = $pdo->prepare($stmtDelete);
+    $stmtDelete = $pdo->prepare($stmtDelete);
 // Execute
-$stmtDelete->execute([":name" => $userToDelete]);
+    $stmtDelete->execute([":name" => $userToDelete]);
+
+}
 
 
 ?>
@@ -50,12 +46,24 @@ $stmtDelete->execute([":name" => $userToDelete]);
           content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" href="style.css">
-    <title>Login</title>
+    <title>Register</title>
 </head>
 <body>
-    <h1>Welcome To The Great Course</h1>
+<?php
+// Show THe Members
+echo "<pre>\n";
+$sqlShow = "SELECT * FROM Usuarios";
+$stmt_show = $pdo->query($sqlShow);
+while($row = $stmt_show->fetch(PDO::FETCH_ASSOC)){
+    echo "<form action='login.php' method='post'><table border='1'><tr><th>ID</th><th>Name</th><th>Email</th></tr><tr><td>{$row["user_id"]}</td> <td>{$row["name"]}</td><td>{$row["email"]}</td></tr></table> <input  type='hidden' name='toDelete' value={$row['user_id']}/> <input type='submit' value='DELETE'></form>";
+}
+echo "</pre>\n";
+?>
 
-    <form method="POST">
+    <h1>Welcome To The Great Course</h1>
+    <h2>Register to Enter </h2>
+
+    <form action="login.php" method="POST">
         <label for="user">Name</label>
         <input type="text" id="user" name="name" required size="40" >
         <br>
@@ -68,7 +76,7 @@ $stmtDelete->execute([":name" => $userToDelete]);
         <input type="submit" value="SEND">
     </form>
 
-
+   <a class="btn" href="./LoginSucces.php">Login</a>
 
 </body>
 </html>
